aider: Added the function get_asana_task that fetches the task gid for the named Asana task in the given project.
# Aider chat conversation:

USER: add the function get_asana_task that is referenced on line 54

ASSISTANT: Sure, I can help with that. We will add the function `get_asana_task` to the `ado_asana_sync.py` file. This function will take two parameters: `asana_project` and `task_name`. It will use the Asana API to fetch tasks from the given project and return the task that matches the given name. If no task is found, it will return `None`.

Here is the code change:

```python
ado_asana_sync.py
<<<<<<< HEAD
=======
def get_asana_task(asana_project, task_name):
    """
    Returns the task gid for the named Asana task in the given project.

    :param asana_project: The gid of the Asana project.
    :type asana_project: str
    :param task_name: The name of the Asana task.
    :type task_name: str
    :return: Task gid or None if no task is found.
    :rtype: str or None
    """
    api_instance = asana.TasksApi(asana_client)
    try:
        # Get all tasks in the project
        api_response = api_instance.get_tasks_in_project(asana_project, completed=False)
        for t in api_response.data:
            if t.name == task_name:
                return t.gid
    except ApiException as e:
        print("Exception when calling TasksApi->get_tasks_in_project: %s\n" % e)
>>>>>>> updated
```

This function uses the `TasksApi` from the Asana client to get all tasks in the given project. It then loops through the tasks and returns the gid of the task that matches the given name. If no task is found, it returns `None`.